requirements.txt
functions-framework==3.*

google.cloud.storage
google.cloud.bigquery
pyarrow

pandas
matplotlib
numpy